Bug 1298801: Includes the daemon command. r=jonasfj
This commits introduces the ability for taskcluster-worker to run as a
daemon. Before running it, you need to install it as a daemon, with the
"install" command:

$ taskcluster-worker daemon install --engine macosx --logging-level info

The --engine argument is mandatory for the install command. To start the
daemon, run:

$ taskcluster-worker daemon start

To stop it:

$ taskcluster-worker daemon stop

And to uninstall it:

$ taskcluster-worker daemon remove

Remember taskcluster-worker reads taskcluster credentials from
environment variables. You need to make sure taskcluster-worker daemon
has these environment variables set.