Enforces correct artifact name. r=jonasfj
osxnative engine creates an on-the-fly user to run its tasks. When the
artifact path is relative, the full path for the artifact is relative
to this temporary user home directory. As the artifact full path may not
have an intersection with artifact name, a string replacement may yield
a invalid artifact name.

This patch uses a concatenation of the artifact directory with artifact
file to generate the full artifact name.