Add configuration for creation to macosx engine.
In order to create some kind of isolation among tasks, macosx engine has
the ability to create an on-fly user to run tasks. This commit adds
configurations to the engine to tell if a new user should be created or
not to run a task, as well as the user group.