Add extra environment variables support.
The child process may expects a set of typical environment variables be
present in the target system. In this patch, we add a configuratio entry
to the env plugin for configuration of extra variables that should be
passed to the sandbox.

We also set the HOME environment variable in the native engine, as many
shell commands expects for it.