Bug 1336585 part 2: Fixes for basic implementation.
* Implement KillChildren function. If we run the command with the same
user of as taskcluster-worker, we can't rely on KillByOwner to kill the
children process spawned in behalf of the task. In this case, we kill
all the children and grandchildren process of the command.

* Set user to nil when not creating a new user. So, if we mistakenly try
to remove the user, taskcluster-worker will panic at worst, but the host
will be safe.

* process_linux and process_darwin become process_posix. The system
process module has exactly the same implementation both in Linux and
Darwin.

* Add more tests.

* Fix bugs and implement reliable process killing.