Bug 1336585 part 3: implement reliable process killing.
Killing the process hierarchy is a tricky operation. First of all, there
is a go routine that waits (man 2 wait) for the parent process. This
means as soon as the parent is killed, *init* will inherit all children
processes. So, we need to kill the children before kill the parent. On
the other hand, there is a risk that after we get the list of children
to kill, the parent creates yet another child process. To solve this
race condition, we first send SIGSTOP to the parent, so the process
will stop and we are sure it won't create any other process. Then we
recursively apply the killProcesses operation to each child, and then
kill the parent.

To actually kill a process, we start sending a SIGTERM, then wait 5
seconds, and then send SIGKILL.