Bug 1339196: Make $HOME and os/user.Lookup().HomeDir equal. r=jonasfj
We can get user's home directory either but looking at the $HOME
environment variable or read the OS specific user database.

As is common for scripts to look for home directory just by reading $HOME
environment variable, our first approach was to not change user database
and modify the $HOME variable to a temporary directory when spawning
the task command without creating a new user. This way, we could clobber
the task just by removing the $HOME directory.

This approach worked for most tests but xpcshell. Particularly, it checks
if $HOME matches user database, making this test to fail.

As we may not be allowed to change even the current user home directory,
we fix this by using the user's true home directory.

This implies that we only need to create a temporary directory if we
create a new user to run the task, otherwise it is not necessary. We
also need to explicitly set the correct owner of the context file.