Added Webhooktunnel support
This commit allows the worker to use webhooktunnel proxy as a
webhookserver implementation to serve livelogs. Webhooktunnel
is a proxy system that allows workers to expose http endpoints
on the proxy instead of opening a public port on the worker.
This results in a more secure worker.
Webhooktunnel works by allowing multiplexing of streams over a
websocket connections. These streams can be used to serve http
requests.
Webhooktunnel requires the tc credentials to work.