Added Webhooktunnel Support
This patch allows the worker to use webhooktunnel to
serve proxy endpoints instead of opening public ports
on the worker.

Webhooktunnel allows http to be served over multiplexed
streams over a websocket. This is exposed as a net.Listener
interface by whclient, and can be used by an http.Server.

Webhooktunnel requires TC credentials for getting an ID and
Token from tc-auth. These can be used to connect to the proxy.