Add get-url support to mockengine as payload function
This is useful during integration tests, as tests can then call
a URL on localhost and let the test framework know how far the
execution have come. This can also be used to block tasks, or simply
check that they run in parallel or sequentially as response will be
printing the task logs.