Add runtime/caching package for implementing caches
The caching package provides an easy way to wrap a constructor such
that objects created are held by the cache and returns things previously
created if available.

Caches can be either shared or exclusive, in case of shared caches at
most one resource will be created for a given input (until disposed),
and if the same resource is requested twice at the same time, it'll only
be created one.

The resources are inserted in garbage collector such that the system
can free resources when it runs low on memory, obviously it uses
reference counting to ensure that resources are not disposed while in
use.