Add tasklog plugin fixing #351 (#352)
This adds a simple tasklog plugin, that just does static logs.
Maybe I should have called it staticlog, but tasklog also works.
There is no reason to run this next to livelog, it's a strict subset
of the features offered by the livelog.

Motivation, ability to better lock down security critical workers,
that you don't want to handle requests, even if they are proxied through
webhooktunnel. Other motivation is that I suspect some of the bugs
@bclary is hitting could be related to livelog or webhooktunnel,
if using this plugin instead of livelog fixes his problems, then
we know for sure we have a bug in livelog. Right now, bugs in livelog
is pure speculation.