Bug 1454401: Implement docker image pull as an artifact
Docker images for Firefox CI are built and made available as a
Taskcluster artifact. To load the image inside a docker environment, we
need to download it and use the "docker load" feature to import it.

We schema for normal image pulling also changed: we now use a single
string to specify the image, and it now supports download by hash.

We remove the image cache as it needs to refactoring to fit wel into
this model.